refactor: Share block volume planning
